Fix clearing scheme
